refactor: 拆分 ChatInput.kt 为多个职责单一的文件
将 1574 行的 ChatInput.kt 按功能拆分为 AsrButton、AttachmentChips、FilesPicker、CropLauncher 四个独立文件
