+++ app/src/main/java/me/rerere/rikkahub/ui/components/ai/AttachmentChips.kt
@@ -1,0 +1,233 @@
+package me.rerere.rikkahub.ui.components.ai
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.horizontalScroll
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.widthIn
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.util.fastForEach
+import androidx.core.net.toUri
+import coil3.compose.AsyncImage
+import me.rerere.ai.ui.UIMessagePart
+import me.rerere.hugeicons.HugeIcons
+import me.rerere.hugeicons.stroke.Cancel01
+import me.rerere.hugeicons.stroke.Files02
+import me.rerere.hugeicons.stroke.MusicNote03
+import me.rerere.hugeicons.stroke.Video01
+import me.rerere.rikkahub.data.files.FilesManager
+import me.rerere.rikkahub.ui.hooks.ChatInputState
+import org.koin.compose.koinInject
+
+@Composable
+internal fun MediaFileInputRow(
+    state: ChatInputState,
+) {
+    val filesManager: FilesManager = koinInject()
+    val managedFiles by filesManager.observe().collectAsState(initial = emptyList())
+    val displayNameByRelativePath = remember(managedFiles) {
+        managedFiles.associate { it.relativePath to it.displayName }
+    }
+    val displayNameByFileName = remember(managedFiles) {
+        managedFiles.associate { it.relativePath.substringAfterLast('/') to it.displayName }
+    }
+
+    fun removePart(part: UIMessagePart, url: String) {
+        state.messageContent = state.messageContent.filterNot { it == part }
+        if (state.shouldDeleteFileOnRemove(part)) {
+            filesManager.deleteChatFiles(listOf(url.toUri()))
+        }
+    }
+
+    Row(
+        horizontalArrangement = Arrangement.spacedBy(8.dp),
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(horizontal = 6.dp, vertical = 6.dp)
+            .horizontalScroll(rememberScrollState())
+    ) {
+        state.messageContent.fastForEach { part ->
+            when (part) {
+                is UIMessagePart.Image -> {
+                    AttachmentChip(
+                        title = attachmentNameFromUrl(
+                            url = part.url,
+                            fallback = "image",
+                            displayNameByRelativePath = displayNameByRelativePath,
+                            displayNameByFileName = displayNameByFileName
+                        ),
+                        leading = {
+                            Surface(
+                                modifier = Modifier.size(34.dp),
+                                shape = RoundedCornerShape(10.dp),
+                                color = MaterialTheme.colorScheme.surfaceContainerHigh,
+                            ) {
+                                AsyncImage(
+                                    model = part.url,
+                                    contentDescription = null,
+                                    contentScale = ContentScale.Crop,
+                                    modifier = Modifier.fillMaxSize()
+                                )
+                            }
+                        },
+                        onRemove = { removePart(part, part.url) }
+                    )
+                }
+
+                is UIMessagePart.Video -> {
+                    AttachmentChip(
+                        title = attachmentNameFromUrl(
+                            url = part.url,
+                            fallback = "video",
+                            displayNameByRelativePath = displayNameByRelativePath,
+                            displayNameByFileName = displayNameByFileName
+                        ),
+                        leading = { AttachmentLeadingIcon(icon = HugeIcons.Video01) },
+                        onRemove = { removePart(part, part.url) }
+                    )
+                }
+
+                is UIMessagePart.Audio -> {
+                    AttachmentChip(
+                        title = attachmentNameFromUrl(
+                            url = part.url,
+                            fallback = "audio",
+                            displayNameByRelativePath = displayNameByRelativePath,
+                            displayNameByFileName = displayNameByFileName
+                        ),
+                        leading = { AttachmentLeadingIcon(icon = HugeIcons.MusicNote03) },
+                        onRemove = { removePart(part, part.url) }
+                    )
+                }
+
+                is UIMessagePart.Document -> {
+                    AttachmentChip(
+                        title = attachmentNameFromUrl(
+                            url = part.url,
+                            fallback = part.fileName,
+                            displayNameByRelativePath = displayNameByRelativePath,
+                            displayNameByFileName = displayNameByFileName
+                        ),
+                        leading = { AttachmentLeadingIcon(icon = HugeIcons.Files02) },
+                        onRemove = { removePart(part, part.url) }
+                    )
+                }
+
+                else -> Unit
+            }
+        }
+    }
+}
+
+@Composable
+private fun AttachmentChip(
+    title: String,
+    leading: @Composable () -> Unit,
+    onRemove: () -> Unit,
+) {
+    Surface(
+        shape = RoundedCornerShape(18.dp),
+        tonalElevation = 1.dp,
+        shadowElevation = 0.dp,
+        color = MaterialTheme.colorScheme.surface,
+        border = BorderStroke(1.dp, MaterialTheme.colorScheme.outlineVariant.copy(alpha = 0.35f))
+    ) {
+        Row(
+            modifier = Modifier
+                .height(44.dp)
+                .padding(start = 8.dp, end = 6.dp),
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.spacedBy(8.dp)
+        ) {
+            leading()
+            Text(
+                text = title,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis,
+                style = MaterialTheme.typography.bodyMedium,
+                modifier = Modifier.widthIn(min = 40.dp, max = 180.dp),
+            )
+            Box(
+                modifier = Modifier
+                    .clip(CircleShape)
+                    .size(26.dp)
+                    .clickable(onClick = onRemove),
+                contentAlignment = Alignment.Center
+            ) {
+                Icon(
+                    imageVector = HugeIcons.Cancel01,
+                    contentDescription = null,
+                    tint = MaterialTheme.colorScheme.onSurfaceVariant,
+                    modifier = Modifier.size(16.dp)
+                )
+            }
+        }
+    }
+}
+
+@Composable
+private fun AttachmentLeadingIcon(
+    icon: ImageVector,
+) {
+    Surface(
+        modifier = Modifier.size(34.dp),
+        shape = RoundedCornerShape(10.dp),
+        color = MaterialTheme.colorScheme.surfaceContainerHigh,
+    ) {
+        Box(
+            modifier = Modifier.fillMaxSize(),
+            contentAlignment = Alignment.Center
+        ) {
+            Icon(
+                imageVector = icon,
+                contentDescription = null,
+                tint = MaterialTheme.colorScheme.onSurfaceVariant
+            )
+        }
+    }
+}
+
+private fun attachmentNameFromUrl(
+    url: String,
+    fallback: String,
+    displayNameByRelativePath: Map<String, String>,
+    displayNameByFileName: Map<String, String>,
+): String {
+    val parsed = runCatching { url.toUri() }.getOrNull()
+    val relativePath = parsed?.path?.substringAfter("/files/", missingDelimiterValue = "")?.takeIf { it.isNotBlank() }
+    if (relativePath != null) {
+        displayNameByRelativePath[relativePath]?.let { return it }
+    }
+
+    val storedFileName = parsed?.lastPathSegment?.substringAfterLast('/')?.takeIf { it.isNotBlank() }
+    if (storedFileName != null) {
+        displayNameByFileName[storedFileName]?.let { return it }
+        return storedFileName
+    }
+
+    return fallback
+}
