+++ app/src/main/java/me/rerere/rikkahub/ui/components/ai/FilesPicker.kt
@@ -1,0 +1,353 @@
+package me.rerere.rikkahub.ui.components.ai
+
+import androidx.compose.foundation.LocalIndication
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.FlowRow
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.wrapContentWidth
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.ListItem
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.ModalBottomSheet
+import androidx.compose.material3.ProvideTextStyle
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.rememberModalBottomSheetState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.semantics.role
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.launch
+import me.rerere.ai.provider.ProviderSetting
+import me.rerere.hugeicons.HugeIcons
+import me.rerere.hugeicons.stroke.Camera01
+import me.rerere.hugeicons.stroke.Files02
+import me.rerere.hugeicons.stroke.Image02
+import me.rerere.hugeicons.stroke.MusicNote03
+import me.rerere.hugeicons.stroke.Package
+import me.rerere.hugeicons.stroke.Package01
+import me.rerere.hugeicons.stroke.Video01
+import me.rerere.rikkahub.R
+import me.rerere.rikkahub.Screen
+import me.rerere.rikkahub.data.datastore.Settings
+import me.rerere.rikkahub.data.datastore.getCurrentChatModel
+import me.rerere.rikkahub.data.datastore.findProvider
+import me.rerere.rikkahub.data.model.Assistant
+import me.rerere.rikkahub.data.model.Conversation
+import me.rerere.rikkahub.ui.components.ui.ExtensionSelector
+import me.rerere.rikkahub.ui.components.ui.permission.PermissionCamera
+import me.rerere.rikkahub.ui.components.ui.permission.PermissionManager
+import me.rerere.rikkahub.ui.components.ui.permission.rememberPermissionState
+import me.rerere.rikkahub.ui.context.LocalNavController
+import me.rerere.rikkahub.ui.context.LocalSettings
+import me.rerere.rikkahub.ui.hooks.ChatInputState
+
+@Composable
+internal fun FilesPicker(
+    conversation: Conversation,
+    assistant: Assistant,
+    state: ChatInputState,
+    onCompressContext: (additionalPrompt: String, targetTokens: Int, keepRecentMessages: Int) -> Job,
+    onUpdateAssistant: (Assistant) -> Unit,
+    showInjectionSheet: Boolean,
+    onShowInjectionSheetChange: (Boolean) -> Unit,
+    showCompressDialog: Boolean,
+    onShowCompressDialogChange: (Boolean) -> Unit,
+    onDismiss: () -> Unit,
+    onTakePic: () -> Unit,
+    onPickImage: () -> Unit,
+    onPickVideo: () -> Unit,
+    onPickAudio: () -> Unit,
+    onPickFile: () -> Unit,
+) {
+    val settings = LocalSettings.current
+    val provider = settings.getCurrentChatModel()?.findProvider(providers = settings.providers)
+
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(16.dp), verticalArrangement = Arrangement.spacedBy(16.dp)
+    ) {
+        FlowRow(
+            modifier = Modifier.fillMaxWidth(),
+            horizontalArrangement = Arrangement.spacedBy(16.dp),
+            verticalArrangement = Arrangement.spacedBy(4.dp),
+        ) {
+            TakePicButton(onLaunchCamera = onTakePic)
+
+            ImagePickButton(onClick = onPickImage)
+
+            if (provider != null && provider is ProviderSetting.Google) {
+                VideoPickButton(onClick = onPickVideo)
+
+                AudioPickButton(onClick = onPickAudio)
+            }
+
+            FilePickButton(onClick = onPickFile)
+        }
+
+        HorizontalDivider(
+            modifier = Modifier.fillMaxWidth()
+        )
+
+        // Extensions (Quick Messages + Prompt Injections + Skills)
+        val activeCount =
+            assistant.quickMessageIds.size + assistant.modeInjectionIds.size + assistant.lorebookIds.size + assistant.enabledSkills.size
+        ListItem(
+            leadingContent = {
+                Icon(
+                    imageVector = HugeIcons.Package,
+                    contentDescription = stringResource(R.string.assistant_page_tab_extensions),
+                )
+            },
+            headlineContent = {
+                Text(stringResource(R.string.assistant_page_tab_extensions))
+            },
+            trailingContent = {
+                if (activeCount > 0) {
+                    Text(
+                        text = activeCount.toString(),
+                        style = MaterialTheme.typography.labelSmall,
+                        color = MaterialTheme.colorScheme.primary,
+                    )
+                }
+            },
+            modifier = Modifier
+                .clip(MaterialTheme.shapes.large)
+                .clickable {
+                    onShowInjectionSheetChange(true)
+                },
+        )
+
+        // Compress History Button
+        ListItem(
+            leadingContent = {
+                Icon(
+                    imageVector = HugeIcons.Package01,
+                    contentDescription = stringResource(R.string.chat_page_compress_context),
+                )
+            },
+            headlineContent = {
+                Text(stringResource(R.string.chat_page_compress_context))
+            },
+            trailingContent = {
+                if (conversation.messageNodes.isNotEmpty()) {
+                    Text(
+                        text = stringResource(R.string.chat_page_message_count, conversation.messageNodes.size),
+                        style = MaterialTheme.typography.labelSmall,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant,
+                    )
+                }
+            },
+            modifier = Modifier
+                .clip(MaterialTheme.shapes.large)
+                .clickable {
+                    onShowCompressDialogChange(true)
+                },
+        )
+    }
+
+    // Injection Bottom Sheet
+    if (showInjectionSheet) {
+        InjectionQuickConfigSheet(
+            assistant = assistant,
+            settings = settings,
+            onUpdateAssistant = onUpdateAssistant,
+            onDismiss = { onShowInjectionSheetChange(false) })
+    }
+
+    // Compress Context Dialog
+    if (showCompressDialog) {
+        CompressContextDialog(onDismiss = {
+            onShowCompressDialogChange(false)
+            onDismiss()
+        }, onConfirm = { additionalPrompt, targetTokens, keepRecentMessages ->
+            onCompressContext(additionalPrompt, targetTokens, keepRecentMessages)
+        })
+    }
+}
+
+@Composable
+private fun InjectionQuickConfigSheet(
+    assistant: Assistant, settings: Settings, onUpdateAssistant: (Assistant) -> Unit, onDismiss: () -> Unit
+) {
+    val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
+    val scope = rememberCoroutineScope()
+    val navController = LocalNavController.current
+
+    ModalBottomSheet(
+        onDismissRequest = onDismiss,
+        sheetState = sheetState,
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .fillMaxHeight(0.75f)
+                .padding(horizontal = 16.dp),
+        ) {
+            ExtensionSelector(
+                assistant = assistant,
+                settings = settings,
+                onUpdate = onUpdateAssistant,
+                modifier = Modifier.weight(1f),
+                onNavigateToQuickMessages = {
+                    scope.launch {
+                        sheetState.hide()
+                        onDismiss()
+                        navController.navigate(Screen.QuickMessages)
+                    }
+                },
+                onNavigateToPrompts = {
+                    scope.launch {
+                        sheetState.hide()
+                        onDismiss()
+                        navController.navigate(Screen.Prompts)
+                    }
+                },
+                onNavigateToSkills = {
+                    scope.launch {
+                        sheetState.hide()
+                        onDismiss()
+                        navController.navigate(Screen.Skills)
+                    }
+                })
+
+            Spacer(modifier = Modifier.height(16.dp))
+        }
+    }
+}
+
+@Composable
+private fun ImagePickButton(onClick: () -> Unit = {}) {
+    BigIconTextButton(icon = {
+        Icon(HugeIcons.Image02, null)
+    }, text = {
+        Text(stringResource(R.string.photo))
+    }) {
+        onClick()
+    }
+}
+
+@Composable
+fun TakePicButton(onLaunchCamera: () -> Unit = {}) {
+    val cameraPermission = rememberPermissionState(PermissionCamera)
+
+    PermissionManager(
+        permissionState = cameraPermission
+    ) {
+        BigIconTextButton(icon = {
+            Icon(HugeIcons.Camera01, null)
+        }, text = {
+            Text(stringResource(R.string.take_picture))
+        }) {
+            if (cameraPermission.allRequiredPermissionsGranted) {
+                onLaunchCamera()
+            } else {
+                cameraPermission.requestPermissions()
+            }
+        }
+    }
+}
+
+@Composable
+fun VideoPickButton(onClick: () -> Unit = {}) {
+    BigIconTextButton(icon = {
+        Icon(HugeIcons.Video01, null)
+    }, text = {
+        Text(stringResource(R.string.video))
+    }) {
+        onClick()
+    }
+}
+
+@Composable
+fun AudioPickButton(onClick: () -> Unit = {}) {
+    BigIconTextButton(icon = {
+        Icon(HugeIcons.MusicNote03, null)
+    }, text = {
+        Text(stringResource(R.string.audio))
+    }) {
+        onClick()
+    }
+}
+
+@Composable
+fun FilePickButton(onClick: () -> Unit = {}) {
+    BigIconTextButton(icon = {
+        Icon(HugeIcons.Files02, null)
+    }, text = {
+        Text(stringResource(R.string.upload_file))
+    }) {
+        onClick()
+    }
+}
+
+@Composable
+private fun BigIconTextButton(
+    modifier: Modifier = Modifier,
+    icon: @Composable () -> Unit,
+    text: @Composable () -> Unit,
+    onClick: () -> Unit,
+) {
+    val interactionSource = remember { MutableInteractionSource() }
+    Column(
+        modifier = modifier
+            .clip(RoundedCornerShape(8.dp))
+            .clickable(
+                interactionSource = interactionSource, indication = LocalIndication.current, onClick = onClick
+            )
+            .semantics {
+                role = Role.Button
+            }
+            .wrapContentWidth(),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.spacedBy(2.dp)) {
+        Surface(
+            tonalElevation = 2.dp, shape = RoundedCornerShape(8.dp)
+        ) {
+            Box(
+                modifier = Modifier.padding(horizontal = 32.dp, vertical = 16.dp)
+            ) {
+                icon()
+            }
+        }
+        ProvideTextStyle(MaterialTheme.typography.bodySmall) {
+            text()
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun BigIconTextButtonPreview() {
+    Row(
+        modifier = Modifier.padding(16.dp)
+    ) {
+        BigIconTextButton(icon = {
+            Icon(HugeIcons.Image02, null)
+        }, text = {
+            Text(stringResource(R.string.photo))
+        }) {}
+    }
+}
